+++ src/Libraries/SmartStore.Data/EfRepository.cs
@@ -23,7 +23,6 @@ namespace SmartStore.Data
         public EfRepository(IDbContext context)
         {
             this._context = context;
-            this.AutoCommitEnabled = true;
         }
 
         #region interface members
@@ -72,8 +71,8 @@ namespace SmartStore.Data
                 throw new ArgumentNullException("entity");
 
             this.Entities.Add(entity);
-			
-            if (this.AutoCommitEnabled)
+
+			if (this.AutoCommitEnabledInternal)
                 _context.SaveChanges();
         }
 
@@ -90,7 +89,7 @@ namespace SmartStore.Data
                     {
                         // insert all in one step
                         entities.Each(x => this.Entities.Add(x));
-                        if (this.AutoCommitEnabled)
+						if (this.AutoCommitEnabledInternal)
                             _context.SaveChanges();
                     }
                     else
@@ -103,7 +102,7 @@ namespace SmartStore.Data
                             saved = false;
                             if (i % batchSize == 0)
                             {
-                                if (this.AutoCommitEnabled)
+								if (this.AutoCommitEnabledInternal)
                                     _context.SaveChanges();
                                 i = 0;
                                 saved = true;
@@ -113,7 +112,7 @@ namespace SmartStore.Data
 
                         if (!saved)
                         {
-                            if (this.AutoCommitEnabled)
+							if (this.AutoCommitEnabledInternal)
                                 _context.SaveChanges();
                         }
                     }
@@ -130,7 +129,7 @@ namespace SmartStore.Data
             if (entity == null)
                 throw new ArgumentNullException("entity");
 
-            if (this.AutoCommitEnabled)
+			if (this.AutoCommitEnabledInternal)
             {
 				if (!InternalContext.Configuration.AutoDetectChangesEnabled)
 				{
@@ -154,7 +153,7 @@ namespace SmartStore.Data
 			if (entities == null)
 				throw new ArgumentNullException("entities");
 
-			if (this.AutoCommitEnabled)
+			if (this.AutoCommitEnabledInternal)
 			{
 				if (!InternalContext.Configuration.AutoDetectChangesEnabled)
 				{
@@ -191,7 +190,7 @@ namespace SmartStore.Data
 			
             this.Entities.Remove(entity);
 
-            if (this.AutoCommitEnabled)
+			if (this.AutoCommitEnabledInternal)
                 _context.SaveChanges();
         }
 
@@ -202,7 +201,7 @@ namespace SmartStore.Data
 
 			this.Entities.RemoveRange(entities);
 
-			if (this.AutoCommitEnabled)
+			if (this.AutoCommitEnabledInternal)
 				_context.SaveChanges();
 		}
 
@@ -249,7 +248,15 @@ namespace SmartStore.Data
             get { return _context; }
         }
 
-        public bool AutoCommitEnabled { get; set; }
+        public bool? AutoCommitEnabled { get; set; }
+
+		private bool AutoCommitEnabledInternal
+		{
+			get
+			{
+				return this.AutoCommitEnabled ?? _context.AutoCommitEnabled;
+			}
+		}
 
         #endregion