+++ src/Libraries/SmartStore.Services/Customers/CustomerService.cs
@@ -434,7 +434,7 @@ namespace SmartStore.Services.Customers
         {
 			var ctx = _customerRepository.Context;
 
-			using (var scope = new DbContextScope(ctx: ctx, autoDetectChanges: false, proxyCreation: true, validateOnSave: false, forceNoTracking: true))
+			using (var scope = new DbContextScope(ctx: ctx, autoDetectChanges: false, proxyCreation: true, validateOnSave: false, forceNoTracking: true, autoCommit: false))
 			{
 				var guestRole = GetCustomerRoleBySystemName(SystemCustomerRoleNames.Guests);
 				if (guestRole == null)
@@ -476,12 +476,7 @@ namespace SmartStore.Services.Customers
 				query = query.OrderBy(c => c.Id);
 
 				var customers = query.Take(maxItemsToDelete).ToList();
-				
-				var crAutoCommit = _customerRepository.AutoCommitEnabled;
-				var gaAutoCommit = _gaRepository.AutoCommitEnabled;
-				_customerRepository.AutoCommitEnabled = false;
-				_gaRepository.AutoCommitEnabled = false;
-				
+
 				int numberOfDeletedCustomers = 0;
 				foreach (var c in customers)
 				{
@@ -494,10 +489,7 @@ namespace SmartStore.Services.Customers
 									  select ga;
 						var attributes = gaQuery.ToList();
 
-						foreach (var attribute in attributes)
-						{
-							_gaRepository.Delete(attribute);
-						}
+						_gaRepository.DeleteRange(attributes);
 						
 						// delete customer
 						_customerRepository.Delete(c);
@@ -524,9 +516,6 @@ namespace SmartStore.Services.Customers
 
 				// save the rest
 				scope.Commit();
-
-				_customerRepository.AutoCommitEnabled = crAutoCommit;
-				_gaRepository.AutoCommitEnabled = gaAutoCommit;
 
 				return numberOfDeletedCustomers;
 			}