+++ src/Libraries/SmartStore.Services/Media/PictureService.cs
@@ -670,14 +670,10 @@ namespace SmartStore.Services.Media
 			var affectedFiles = new List<string>(1000);
 
 			var ctx = _pictureRepository.Context;
-
-			_pictureRepository.AutoCommitEnabled = false;
-
 			var failed = false;
-
 			int i = 0;
 
-			using (var scope = new DbContextScope(ctx: ctx, autoDetectChanges: false, proxyCreation: false, validateOnSave: false))
+			using (var scope = new DbContextScope(ctx: ctx, autoDetectChanges: false, proxyCreation: false, validateOnSave: false, autoCommit: false))
 			{
 				using (var tx = ctx.BeginTransaction())
 				{
@@ -692,7 +688,7 @@ namespace SmartStore.Services.Media
 							if (pictures != null)
 							{
 								// detach all entities from previous page to save memory
-								pictures.Each(x => ctx.Detach(x));
+								pictures.Each(x => ctx.DetachEntity(x));
 
 								// breathe
 								pictures.Clear();
@@ -759,8 +755,6 @@ namespace SmartStore.Services.Media
 					}
 				}		
 			}
-
-			_pictureRepository.AutoCommitEnabled = true;
 
 			if (affectedFiles.Count > 0)
 			{