+++ src/Libraries/SmartStore.Services/Tasks/ITaskExecutor.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using SmartStore.Core.Domain.Tasks;
+
+namespace SmartStore.Services.Tasks
+{
+    public interface ITaskExecutor
+    {
+        void Execute(ScheduleTask task, bool throwOnError = false);
+    }
+}