+++ src/Presentation/SmartStore.Web/Administration/Controllers/ScheduleTaskController.cs
@@ -21,6 +21,7 @@ using SmartStore.Core.Async;
 using Autofac;
 using SmartStore.Core.Logging;
 using SmartStore.Core.Plugins;
+using System.Collections.Generic;
 
 namespace SmartStore.Admin.Controllers
 {
@@ -30,6 +31,7 @@ namespace SmartStore.Admin.Controllers
         #region Fields
 
 		private readonly IScheduleTaskService _scheduleTaskService;
+        private readonly ITaskScheduler _taskScheduler;
         private readonly IPermissionService _permissionService;
         private readonly IDateTimeHelper _dateTimeHelper;
 
@@ -37,9 +39,14 @@ namespace SmartStore.Admin.Controllers
 
         #region Constructors
 
-		public ScheduleTaskController(IScheduleTaskService scheduleTaskService, IPermissionService permissionService, IDateTimeHelper dateTimeHelper)
+		public ScheduleTaskController(
+            IScheduleTaskService scheduleTaskService, 
+            ITaskScheduler taskScheduler, 
+            IPermissionService permissionService, 
+            IDateTimeHelper dateTimeHelper)
         {
             this._scheduleTaskService = scheduleTaskService;
+			this._taskScheduler = taskScheduler;
             this._permissionService = permissionService;
             this._dateTimeHelper = dateTimeHelper;
         }
@@ -48,8 +55,11 @@ namespace SmartStore.Admin.Controllers
 
         #region Utility
 
-		private bool IsTaskInstalled(ScheduleTask task)
+		private bool IsTaskVisible(ScheduleTask task)
 		{
+			if (task.IsHidden)
+				return false;
+
 			var type = Type.GetType(task.Type);
 			if (type != null)
 			{
@@ -61,25 +71,55 @@ namespace SmartStore.Admin.Controllers
         [NonAction]
         protected ScheduleTaskModel PrepareScheduleTaskModel(ScheduleTask task)
         {
-            var model = new ScheduleTaskModel()
-            {
-                Id = task.Id,
-                Name = task.Name,
-                Seconds = task.Seconds,
-                Enabled = task.Enabled,
-                StopOnError = task.StopOnError,
-                LastStartUtc = task.LastStartUtc.HasValue ? _dateTimeHelper.ConvertToUserTime(task.LastStartUtc.Value, DateTimeKind.Utc).ToString("G") : "",
-                LastEndUtc = task.LastEndUtc.HasValue ? _dateTimeHelper.ConvertToUserTime(task.LastEndUtc.Value, DateTimeKind.Utc).ToString("G") : "",
-                LastSuccessUtc = task.LastSuccessUtc.HasValue ? _dateTimeHelper.ConvertToUserTime(task.LastSuccessUtc.Value, DateTimeKind.Utc).ToString("G") : "",
+			var now = DateTime.UtcNow;
+			
+			TimeSpan? dueIn = null;
+			if (task.NextRunUtc.HasValue)
+			{
+				dueIn = task.NextRunUtc.Value - now;
+			}
+
+			var nextRunStr = "";
+			bool isOverdue = false;
+			if (dueIn.HasValue)
+			{
+				if (dueIn.Value.TotalSeconds > 0)
+				{
+					nextRunStr = "<span class='muted'>{0}</span>".FormatCurrent(dueIn.Value.Prettify());
+				}
+				else
+				{
+					nextRunStr = "<span class='text-success'><strong>{0}</strong></span>".FormatCurrent(T("Common.Waiting") + "...");
+					isOverdue = true;
+				}
+			}
+
+			var isRunning = task.IsRunning;
+
+			var model = new ScheduleTaskModel
+			{
+				Id = task.Id,
+				Name = task.Name,
+				Seconds = task.Seconds,
+				Enabled = task.Enabled,
+				StopOnError = task.StopOnError,
+				LastStartUtc = task.LastStartUtc.HasValue ? task.LastStartUtc.Value.RelativeFormat(true, "f") : "",
+				LastEndUtc = task.LastEndUtc.HasValue ? _dateTimeHelper.ConvertToUserTime(task.LastEndUtc.Value, DateTimeKind.Utc).ToString("G") : "",
+				LastSuccessUtc = task.LastSuccessUtc.HasValue ? _dateTimeHelper.ConvertToUserTime(task.LastSuccessUtc.Value, DateTimeKind.Utc).ToString("G") : "",
+				NextRunUtc = nextRunStr,
 				LastError = task.LastError.EmptyNull(),
-				IsRunning =	task.IsRunning,		//task.LastStartUtc.GetValueOrDefault() > task.LastEndUtc.GetValueOrDefault(),
+				IsRunning = isRunning,
+				CancelUrl = isRunning ? Url.Action("CancelJob", new { id = task.Id }) : "",
+				ProgressPercent = task.ProgressPercent,
+				ProgressMessage = task.ProgressMessage,
+				IsOverdue = isOverdue,
 				Duration = ""
-            };
+			};
 
 			var span = TimeSpan.Zero;
 			if (task.LastStartUtc.HasValue)
 			{
-				span = model.IsRunning ? DateTime.UtcNow - task.LastStartUtc.Value : task.LastEndUtc.Value - task.LastStartUtc.Value;
+				span = model.IsRunning ? now - task.LastStartUtc.Value : task.LastEndUtc.Value - task.LastStartUtc.Value;
 				model.Duration = span.ToString("g");
 			}
 
@@ -110,9 +150,13 @@ namespace SmartStore.Admin.Controllers
                 return AccessDeniedView();
 			
             var models = _scheduleTaskService.GetAllTasks(true)
-				.Where(IsTaskInstalled)
+				.Where(IsTaskVisible)
                 .Select(PrepareScheduleTaskModel)
+				.OrderByDescending(x => x.IsRunning)
+				.ThenByDescending(x => x.IsOverdue)
+				.ThenBy(x => x.Seconds)
                 .ToList();
+
             var model = new GridModel<ScheduleTaskModel>
             {
                 Data = models,
@@ -135,17 +179,36 @@ namespace SmartStore.Admin.Controllers
             {
                 //display the first model error
                 var modelStateErrors = this.ModelState.Values.SelectMany(x => x.Errors).Select(x => x.ErrorMessage);
+				NotifyError(modelStateErrors.FirstOrDefault());
                 return Content(modelStateErrors.FirstOrDefault());
             }
 
             var scheduleTask = _scheduleTaskService.GetTaskById(model.Id);
-            if (scheduleTask == null)
-                return Content("Schedule task cannot be loaded");
+			if (scheduleTask == null)
+			{
+				NotifyError("Schedule task cannot be loaded");
+				return Content("");
+			}
+
+			if (scheduleTask.IsRunning)
+			{
+				NotifyError(T("Admin.System.ScheduleTasks.UpdateLocked"));
+				return Content("");
+			}
 
             scheduleTask.Name = model.Name;
             scheduleTask.Seconds = model.Seconds;
             scheduleTask.Enabled = model.Enabled;
             scheduleTask.StopOnError = model.StopOnError;
+
+			if (model.Enabled)
+			{
+				scheduleTask.NextRunUtc = DateTime.UtcNow.AddSeconds(scheduleTask.Seconds);
+			}
+			else
+			{
+				scheduleTask.NextRunUtc = null;
+			}
 
 			int max = Int32.MaxValue / 1000;
 
@@ -156,6 +219,23 @@ namespace SmartStore.Admin.Controllers
             return List(command);
         }
 
+		[HttpPost]
+		public ActionResult GetRunningTasks()
+		{
+			if (!_scheduleTaskService.HasRunningTasks())
+				return Json(null);
+
+			var runningTasks = from t in _scheduleTaskService.GetRunningTasks()
+							   select new 
+							   {
+ 								   id = t.Id,
+								   percent = t.ProgressPercent,
+								   message = t.ProgressMessage,
+							   };
+
+			return Json(runningTasks.ToArray());
+		}
+
 		public ActionResult RunJob(int id, string returnUrl = "")
 		{
 			if (!_permissionService.Authorize(StandardPermissionProvider.ManageScheduleTasks))
@@ -163,45 +243,25 @@ namespace SmartStore.Admin.Controllers
 
 			returnUrl = returnUrl.NullEmpty() ?? Request.UrlReferrer.ToString();
 
-			var t = AsyncRunner.Run((c, ct) =>
+            _taskScheduler.RunSingleTask(id);
+
+            NotifyInfo(T("Admin.System.ScheduleTasks.RunNow.Progress"));
+			return Redirect(returnUrl);
+		}
+
+		public ActionResult CancelJob(int id /* scheduleTaskId */, string returnUrl = "")
+		{
+			if (!_permissionService.Authorize(StandardPermissionProvider.ManageScheduleTasks))
+				return AccessDeniedView();
+
+			var cts = _taskScheduler.GetCancelTokenSourceFor(id);
+			if (cts != null)
 			{
-				var svc = c.Resolve<IScheduleTaskService>();
-
-				try
-				{
-					var scheduleTask = svc.GetTaskById(id);
-					if (scheduleTask == null)
-						throw new Exception("Schedule task cannot be loaded");
-
-					var job = new Job(scheduleTask);
-					job.Enabled = true;
-					job.Execute(ct, c, false);
-				}
-				catch
-				{
-					svc.EnsureTaskIsNotRunning(id);
-				}
-			});
-
-			// wait only 100 ms.
-			t.Wait(100);
-
-			if (t.IsCompleted)
-			{
-				if (!t.IsFaulted)
-				{
-					NotifySuccess(T("Admin.System.ScheduleTasks.RunNow.Completed"));
-				}
-				else
-				{
-					NotifyError(t.Exception.Flatten().InnerException);
-				}
-			}
-			else
-			{
-				NotifyInfo(T("Admin.System.ScheduleTasks.RunNow.Progress"));
+				cts.Cancel();
+				NotifyWarning(T("Admin.System.ScheduleTasks.CancellationRequested"));
 			}
 
+			returnUrl = returnUrl.NullEmpty() ?? Request.UrlReferrer.ToString();
 			return Redirect(returnUrl);
 		}