+++ src/Presentation/SmartStore.Web/Administration/Models/Payments/PaymentMethodEditModel.cs
@@ -2,12 +2,27 @@
 using System.Web.Mvc;
 using SmartStore.Core.Domain.Common;
 using SmartStore.Web.Framework;
+using SmartStore.Web.Framework.Localization;
 using SmartStore.Web.Framework.Mvc;
 
 namespace SmartStore.Admin.Models.Payments
 {
-	public class PaymentMethodEditModel : ProviderModel
+	public class PaymentMethodEditModel : EntityModelBase, ILocalizedModel<PaymentMethodLocalizedModel>
 	{
+		public PaymentMethodEditModel()
+		{
+			Locales = new List<PaymentMethodLocalizedModel>();
+		}
+
+		public IList<PaymentMethodLocalizedModel> Locales { get; set; }
+		public string IconUrl { get; set; }
+
+		[SmartResourceDisplayName("Common.SystemName")]
+		public string SystemName { get; set; }
+
+		[SmartResourceDisplayName("Common.FriendlyName")]
+		public string FriendlyName { get; set; }
+
 		[SmartResourceDisplayName("Admin.Configuration.Payment.Methods.ExcludedCustomerRole")]
 		public string[] ExcludedCustomerRoleIds { get; set; }
 		public List<SelectListItem> AvailableCustomerRoles { get; set; }
@@ -33,5 +48,30 @@ namespace SmartStore.Admin.Models.Payments
 		[SmartResourceDisplayName("Admin.Configuration.Restrictions.AmountRestrictionContext")]
 		public AmountRestrictionContextType AmountRestrictionContext { get; set; }
 		public List<SelectListItem> AvailableAmountRestrictionContextTypes { get; set; }
+
+		[SmartResourceDisplayName("Admin.Configuration.Payment.Methods.ShortDescription")]
+		[AllowHtml]
+		public string Description { get; set; }
+
+		[SmartResourceDisplayName("Admin.Configuration.Payment.Methods.FullDescription")]
+		[AllowHtml]
+		public string FullDescription { get; set; }
+	}
+
+
+	public class PaymentMethodLocalizedModel : ILocalizedModelLocal
+	{
+		public int LanguageId { get; set; }
+
+		[SmartResourceDisplayName("Common.FriendlyName")]
+		public string FriendlyName { get; set; }
+
+		[SmartResourceDisplayName("Admin.Configuration.Payment.Methods.ShortDescription")]
+		[AllowHtml]
+		public string Description { get; set; }
+
+		[SmartResourceDisplayName("Admin.Configuration.Payment.Methods.FullDescription")]
+		[AllowHtml]
+		public string FullDescription { get; set; }
 	}
 }